requirements.txt
pyyaml
mysql-connector-python
mock
pytest
coverage
